Use options for case insensitivity and extra infos
Word list should be in all lowercase. In case it is all uppercase, the
switch `-i` makes all underlying `grep` not case sensitive. Beware,
there are some gliches with direct use of system words list: there is
a bulk of proper nouns that start with capital letters and the script
don't deal with that.

Summary of parameters used are no longer displayed unless `-d` is
supplied. That makes the script more suitable for piping its results.